feat: editing common fields

+++ backend/app/app/crud/base.py
@@ -61,6 +61,9 @@ class CRUDBase(Generic[ModelType, CreateSchemaType, UpdateSchemaType]):
 
     def remove(self, db: Session, *, id: int) -> ModelType:
         obj = db.query(self.model).get(id)
-        db.delete(obj)
-        db.commit()
+
+        if obj:
+            db.delete(obj)
+            db.commit()
+        
         return obj
